backend/requirements.txt
fastapi[standard]
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
pgvector
pydantic-settings
httpx
apscheduler
supabase
google-generativeai
sentence-transformers
numpy
python-dotenv
